Extract the resolver into its own library
