+++ src/generator/include/sourcemeta/registry/generator.h
@@ -1,10 +1,8 @@
 #ifndef SOURCEMETA_REGISTRY_GENERATOR_H_
 #define SOURCEMETA_REGISTRY_GENERATOR_H_
 
-#include <sourcemeta/registry/generator_collection.h>
 #include <sourcemeta/registry/generator_configuration.h>
 #include <sourcemeta/registry/generator_output.h>
-#include <sourcemeta/registry/generator_resolver.h>
 #include <sourcemeta/registry/generator_validator.h>
 
 #endif
