+++ test/unit/resolver/registry.json
