+++ test/unit/resolver/schemas/example/2019-09-recursive-ref.json
