+++ test/unit/resolver/schemas/example/2020-12-embedded-resource.json
