+++ test/unit/resolver/schemas/example/2020-12-id-with-casing.json
