+++ test/unit/resolver/schemas/example/2020-12-ref-needs-rebase.json
