+++ test/unit/resolver/schemas/example/draft4-internal-ref.json
