+++ test/unit/resolver/schemas/example/draft4-trailing-hash-with-ref.json
