+++ test/unit/resolver/schemas/example/only-id.json
