+++ test/unit/resolver/schemas/meta-draft4/draft4-override.json
